requirements.txt
Flask
cryptography
PyJWT
pytest     # For testing
